pyproject.toml
[tool.poetry]
name = "repl-nix-thesis-website-project"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "repl_nix_thesis_website_project"}]

[tool.poetry.dependencies]
python = "^3.10"
flask = "^3.0.3"
markupsafe = "^2.1.5"
paho-mqtt = "^2.1.0"
pyrebase = "^3.0.27"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
